fix&add いいねボタンの修正とリツイート機能の追加

+++ likes.php
@@ -2,21 +2,21 @@
 session_start();
 require('dbconnect.php');
 
-if (isset($_SESSION['id'])) {
-    $id = $_SESSION['id']; //member_id
-    $post = $_REQUEST['id']; //post_id
+if (isset($_SESSION['id']) && isset($_REQUEST['id'])) {
+    $member_id = $_SESSION['id'];
+    $post_id = $_REQUEST['id'];
 
     $check_likes = $db->prepare('SELECT COUNT(*) AS count FROM likes WHERE member_id = ? and post_id = ?');
-    $check_likes ->execute(array($id,$post));
+    $check_likes ->execute(array($member_id,$post_id));
     $check = $check_likes->fetch();
 
     //イイねが0ならデータを追加、0以外なら削除
     if ($check['count'] == 0) {
         $add_likes = $db->prepare('INSERT INTO likes SET member_id=?, post_id=?');
-        $add_likes ->execute(array($id,$post));
+        $add_likes ->execute(array($member_id,$post_id));
     } else {
         $del_likes = $db->prepare('DELETE FROM likes WHERE member_id=? AND post_id=?');
-        $del_likes ->execute(array($id,$post));
+        $del_likes ->execute(array($member_id,$post_id));
     }
 }
